backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
coverage==7.6.10
dj-rest-auth==7.0.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==33.3.1
flake8==7.1.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
